+++ lib/hnco/functions/walsh/walsh-expansion-1.hh
@@ -120,7 +120,7 @@ namespace function {
     ///@{
 
     /// Get bit vector size
-    size_t get_bv_size() { return _linear.size(); }
+    int get_bv_size() { return _linear.size(); }
 
     /// Get the global maximum
     double get_maximum();
